requirements.txt
asyncio
scrapy
pandas
openpyxl
aiogram
requests
bs4
lxml
xlsxwriter
schedule
playwright
fake-useragent
python-dotenv
